Task A
I made html first, then I start to setting up what was asking for Task A. I created database in MySQL called clients_form with mysql_table. Column names as snake_case. After that, I save the excel as Client.csv, and I made index.js to read the "excel", to check if the data in the CSV is valid. With, error message if failed. Finally I made too the database.js, to connects to MYSQL.